Add edge casing to setup script, add run script for ease of use

+++ setup.sh
@@ -2,18 +2,26 @@
 RED='\033[0;31m'
 GREEN='\033[0;32m'
 YELLOW='\033[0;33m'
-BLUE='\033[0;34m'
+BLUE='\033[0;36m'
+NC='\033[0m'
 
 
 gpio_txt=$( cat res/gpio_script.txt )
+dir="src/RPi"
 
 if [ "$1" = "-w"  ]; then
 	printf "Are you sure you want to setup in Testing / Devlopment mode? (y/n): "
 	read ans
 	if [ "$ans" = "y" ]; then
-		printf "${YELLOW}Setting up in Testing / Development Mode\n"
-		mkdir  src/RPi
+		printf "${BLUE}Setting up in Testing / Development Mode\n"
+		if [[ -d $dir ]]; then
+			printf "${YELLOW}Already in Development mode ($dir already exists). Cancelling...\n"
+			exit 0
+		fi
+		mkdir $dir
+		printf "${NC}Creating src/RPi directory\n"
 		touch src/RPi/__init__.py src/RPi/GPIO.py
+		printf "Creating files for dev environment\n"
 		echo -e "$gpio_txt" > src/RPi/GPIO.py
 		printf "${GREEN}Succesfully Setup Testing / Development Environment\n"
 	elif [ "$ans" = "n"  ]; then
@@ -35,6 +43,10 @@ elif [ "$1" = "--help" ]; then
 	exit 0
 elif [ "$1" = "-c" ]; then
 	printf "Cleaning up...\n"
+	if [[ ! -d $dir ]]; then
+		printf "${YELLOW}Already ready for Desployment. Cancelling...\n"
+		exit 0
+	fi
 	rm -r "src/RPi"
 	printf "${GREEN}Setup success, ready for deployment.\n"
 	exit 0
